final bunny bot code

+++ bunny2018/shooter/ShooterEjectAutoBallAction.h
@@ -25,7 +25,7 @@ namespace xero {
             virtual bool ejectTrigger() ;
             virtual bool repeatEjectOne() ;
 
-			double getClosestCrate(double pos, bool fwd) ;
+			size_t getClosestCrate(double pos, bool fwd) ;
 
         private:
             std::shared_ptr<xero::base::DriveBase> db_ ;
